functions seperated and norm checked, malloc checks and exits for them are  still missing
