+++ includes/cub3d.h
@@ -14,6 +14,7 @@
 # include <unistd.h>
 # include <fcntl.h>
 # include <math.h>
+# include <stdio.h> //remove;
 
 typedef struct s_fcheck
 {
@@ -114,5 +115,34 @@ int		wallfloodfill(t_map *map, int posy, int posx);
 int		check_num(char c);
 void	loop_trough(t_map *map, char *str, int count);
 
+
+//parsing files
+
+void	check_int_of_rgb(t_fcheck *check, char **ceiling);
+void	check_rgb_floor(t_fcheck *check);
+void	check_rgb_ceiling(t_fcheck *check);
+void	looptrough(t_fcheck *check, char *str, int count);
+void	direction_check(t_fcheck *check, int *invalid, int *i, int *j);
+int 	check_char(char c);
+void	map_boarder_check(t_fcheck *check);
+void	check_player_and_boarder(t_fcheck *check);
+int		extract_and_set_info1(char *str, int info_set, t_fcheck *check);
+int		extract_and_set_info2(char *str, int info_set, t_fcheck *check);
+char	*check_and_set_info(char *str, char *info, t_fcheck *check);
+void	check_info_flag(t_fcheck *check, int i);
+void	check_failed_info_flag(t_fcheck *check);
+void	invalid_input(char *str);
+void	error_inside_file(t_fcheck *check);
+void	clean_from_space_nl(char **str);
+void	clean_from_nl(char **str);
+void	clean_cardinal_directions(char **str);
+void	clean_input_strings(t_fcheck *check);
+int		count_file_lines(char *map_name, int count);
+void	read_file(char *map_name, int count);
+void	get_map_information(t_fcheck *check);
+int		check_rest(t_fcheck *check, int	i);
+void	argument_check(char *argv);
+// remove;
+void	print_check_struct(t_fcheck *check);
 
 #endif
