+++ map3.cub
@@ -22,7 +22,7 @@ C 225,        30,   0
                    100000000001
                    100000000001
 1111111111111111111100000000001
-1000000000000000000000200000001
+1000000000000000000000000000001
 1000000000000000000000000000001
 1000000000000000000000000000001
 1000000000001111111111111111111
